queue agent messages instead of directly sending
As there is a limit to how many/fast agent messages may be sent
controlled by the server tokens we need to queue the messages and send
them as we are given more tokens by the server. This fixes the old sleep
behaviour preventing deadlocks when the calling library calls clipboard
functions from the same thread as the spice_process thread.