Use channel specific connect packets
The channel connect packet size & members is defined by the number of
value in the enumerations and as such can be different in size.
Currently this is not the case but should the spice protocol be enhanced
with additional enumerations in the future it's better to be prepared
for this.